# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Micro Task Template API
  version: 1.0.0
  description: >
    API Gateway with Users and Orders services. Versioned under /v1.
servers:
  - url: http://localhost:8000
    description: Local via API Gateway
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequestIdHeader:
      in: header
      name: X-Request-ID
      required: false
      schema: { type: string }
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        error:
          nullable: true
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        roles:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OrderItem:
      type: object
      properties:
        sku: { type: string }
        name: { type: string }
        quantity: { type: integer }
        price: { type: number }
      required: [sku, name, quantity, price]
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        status:
          type: string
          enum: [created, in_progress, completed, cancelled]
        totalAmount: { type: number }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
paths:
  /v1/users/register:
    post:
      summary: Register new user
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                name: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '409':
          description: Already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /v1/users/login:
    post:
      summary: Login and receive JWT
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /v1/users/me:
    get:
      summary: Get current profile
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    put:
      summary: Update current profile
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                roles:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /v1/users:
    get:
      summary: List users (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
        - in: query
          name: role
          schema: { type: string }
        - in: query
          name: name
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /v1/orders:
    post:
      summary: Create order
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items, totalAmount]
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/OrderItem' }
                totalAmount: { type: number }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    get:
      summary: List orders of current user (admin/manager see all)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
        - in: query
          name: sort
          schema: { type: string, default: "createdAt:desc" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /v1/orders/{orderId}:
    get:
      summary: Get order by id
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - in: path
          name: orderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    put:
      summary: Update order status (manager/admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - in: path
          name: orderId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [created, in_progress, completed, cancelled]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '400':
          description: Invalid transition
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /v1/orders/{orderId}/cancel:
    post:
      summary: Cancel own order
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - in: path
          name: orderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }